backend/pyproject.toml
[project]
name = "paylens-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "google-genai>=1.68.0",
    "httpx>=0.28.1",
    "litellm>=1.82.4",
    "pydantic>=2.0",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.42.0",
]
